enable full document restyling

+++ src/Parser.cpp
@@ -94,14 +94,14 @@ int *ParseFile(tinyxml2::XMLElement *config, int length, std::string text)
             for (; it != end; ++it)
             {
                 boost::smatch m = *it;
-                int numGroups = m.size();
-                int strLen = m.length();
-                int pos = m.position();
+                unsigned long long numGroups = m.size();
+                long long strLen = m.length();
+                long long pos = m.position();
                 if (numGroups > 1) // regex contains groups, need to parse each group and stylize them appropriately
                 {
-                    for (int j = 1; j <= numGroups; j++)
+                    for (unsigned long long j = 1; j <= numGroups; j++)
                     {
-                        for (int k = 1; k <= m[j].length(); k++)
+                        for (long long k = 1; k <= m[j].length(); k++)
                         {
                             styles[m.position(j) + k] = token.style;
                         }
@@ -109,7 +109,7 @@ int *ParseFile(tinyxml2::XMLElement *config, int length, std::string text)
                 }
                 else // regex contains no groups, just stylize the matched string
                 {
-                    for (int j = 1; j <= strLen; j++)
+                    for (long long j = 1; j <= strLen; j++)
                     {
                         styles[pos + j] = token.style;
                     }
